+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog]
 
 ## [Unreleased]
+- Google Analytics
+  - Adds CSV view for claim stats
 
 ## [Release 095] - 2021-09-09